+++ Sources/EmceeLib/Commands/DumpRuntimeTestsCommand.swift
@@ -59,6 +59,7 @@ public final class DumpRuntimeTestsCommand: Command {
             let configuration = RuntimeDumpConfiguration(
                 developerDir: testArgFileEntry.toolchainConfiguration.developerDir,
                 runtimeDumpMode: try RuntimeDumpModeDeterminer.runtimeDumpMode(testArgFileEntry: testArgFileEntry),
+                simulatorSettings: testArgFileEntry.simulatorSettings,
                 testDestination: testArgFileEntry.testDestination,
                 testExecutionBehavior: TestExecutionBehavior(
                     environment: testArgFileEntry.environment,