+++ Sources/Models/ResourceAliases.swift
@@ -10,7 +10,6 @@ public typealias FbsimctlLocation = TypedResourceLocation<FbsimctlResourceLocati
 public typealias FbxctestLocation = TypedResourceLocation<FbxctestResourceLocationType>
 public typealias PluginLocation = TypedResourceLocation<PluginResourceLocationType>
 public typealias RunnerAppLocation = TypedResourceLocation<RunnerAppResourceLocationType>
-public typealias SimulatorLocalizationLocation = TypedResourceLocation<SimulatorLocalizationResourceLocationType>
 public typealias TestBundleLocation = TypedResourceLocation<TestBundleResourceLocationType>
 public typealias QueueServerRunConfigurationLocation = TypedResourceLocation<QueueServerRunConfigurationLocationType>
 public typealias AnalyticsConfigurationLocation = TypedResourceLocation<AnalyticsConfigurationLocationType>
@@ -38,10 +37,6 @@ public final class PluginResourceLocationType: ResourceLocationType {
 
 public final class RunnerAppResourceLocationType: ResourceLocationType {
     public static let name = "xct runner app"
-}
-
-public final class SimulatorLocalizationResourceLocationType: ResourceLocationType {
-    public static let name = "simulator localization settings"
 }
 
 public final class TestBundleResourceLocationType: ResourceLocationType {