+++ Sources/RuntimeDump/RuntimeTestQuerierImpl.swift
@@ -118,10 +118,6 @@ public final class RuntimeTestQuerierImpl: RuntimeTestQuerier {
         dumpConfiguration: RuntimeDumpConfiguration,
         runtimeEntriesJSONPath: AbsolutePath
     ) -> RunnerConfiguration {
-        let simulatorSettings = SimulatorSettings(
-            simulatorLocalizationSettings: nil,
-            watchdogSettings: WatchdogSettings(bundleIds: [], timeout: 0)
-        )
         let environment = self.environment(configuration: dumpConfiguration, runtimeEntriesJSONPath: runtimeEntriesJSONPath)
 
         switch dumpConfiguration.runtimeDumpMode {
@@ -134,7 +130,7 @@ public final class RuntimeTestQuerierImpl: RuntimeTestQuerier {
                     )
                 ),
                 environment: environment,
-                simulatorSettings: simulatorSettings,
+                simulatorSettings: dumpConfiguration.simulatorSettings,
                 testRunnerTool: dumpConfiguration.testRunnerTool,
                 testTimeoutConfiguration: dumpConfiguration.testTimeoutConfiguration,
                 testType: .logicTest
@@ -151,7 +147,7 @@ public final class RuntimeTestQuerierImpl: RuntimeTestQuerier {
                     additionalApplicationBundles: []
                 ),
                 environment: environment,
-                simulatorSettings: simulatorSettings,
+                simulatorSettings: dumpConfiguration.simulatorSettings,
                 testRunnerTool: dumpConfiguration.testRunnerTool,
                 testTimeoutConfiguration: dumpConfiguration.testTimeoutConfiguration,
                 testType: .appTest